rename db <- db-config
